Move NaiveTreeSHAP to h2o-algos:test to make it available to GBM tests
