+++ h2o-extensions/xgboost/build.gradle
@@ -10,6 +10,7 @@ dependencies {
     testCompile "junit:junit:${junitVersion}"
     testCompile 'com.esotericsoftware.kryo:kryo:2.21'
     testCompile project(path: ":h2o-core", configuration: "testArchives")
+    testCompile project(path: ":h2o-algos", configuration: "testArchives")
     testRuntimeOnly project(":${defaultWebserverModule}")
 }
